+++ packages/frontend-sdk/index.ts
@@ -1,0 +1,2 @@
+export {GrabberPlayerClient} from "./lib/grabber_player"
+export {GrabberSocket} from "./lib/sockets"
