--- packages/frontend-sdk/lib/main.ts
@@ -1,2 +1,0 @@
-export {GrabberPlayerClient} from "./grabber_player"
-export {GrabberSocket} from "./sockets"
